+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to the full browser extension will be documented in this fil
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Umreleased
+### Added
+- Report `textarea` and `select` elements.
+
 ## 0.1.9 - 2026-05-15
 
 ### Added